+++ src/bin-utils.js
@@ -76,6 +76,13 @@ export const initSegmentId = function(initSegment) {
 };
 
 /**
+ * Returns a unique string identifier for a media segment key.
+ */
+export const segmentKeyId = function(key) {
+  return key.resolvedUri;
+};
+
+/**
  * utils to help dump binary data to the console
  */
 export const hexDump = (data) => {