+++ test/source-updater.test.js
@@ -131,6 +131,24 @@ QUnit.test('runs the next callback after updateend fires', function(assert) {
                   'appended the bytes');
 });
 
+QUnit.test('runs the next callback after calling timestampOffset', function(assert) {
+  let updater = new SourceUpdater(this.mediaSource, 'video/mp2t');
+  let sourceBuffer;
+
+  updater.timestampOffset(10);
+  updater.appendBuffer({
+    bytes: new Uint8Array([0, 1, 2])
+  }, () => {});
+
+  this.mediaSource.trigger('sourceopen');
+  sourceBuffer = this.mediaSource.sourceBuffers[0];
+
+  assert.equal(sourceBuffer.timestampOffset, 10, 'offset correctly set');
+  assert.equal(sourceBuffer.updates_.length, 1, 'updated once');
+  assert.deepEqual(sourceBuffer.updates_[0].append, new Uint8Array([0, 1, 2]),
+                  'appended the bytes');
+});
+
 QUnit.test('runs only one callback at a time', function(assert) {
   let updater = new SourceUpdater(this.mediaSource, 'video/mp2t');
   let sourceBuffer;
@@ -259,4 +277,26 @@ QUnit.test('supports timestampOffset', function(assert) {
 
   sourceBuffer.trigger('updateend');
   assert.equal(sourceBuffer.timestampOffset, 14, 'applied the update');
+});
+
+QUnit.test('abort on dispose waits until after a remove has finished', function(assert) {
+  let updater = new SourceUpdater(this.mediaSource, 'video/mp2t');
+  let sourceBuffer;
+
+  this.mediaSource.trigger('sourceopen');
+  updater.appendBuffer({
+    bytes: new Uint8Array([0])
+  }, () => {});
+
+  sourceBuffer = this.mediaSource.sourceBuffers[0];
+  sourceBuffer.trigger('updateend');
+  updater.remove(0, 10);
+  updater.dispose();
+
+  assert.deepEqual(sourceBuffer.updates_[1].remove, [0, 10], 'remove called');
+  assert.equal(sourceBuffer.updates_.length, 2, 'abort not called before updateend');
+
+  sourceBuffer.trigger('updateend');
+
+  assert.ok(sourceBuffer.updates_[2].abort, 'aborted the source buffer');
 });